Zona radiactiva por los cuatro lados: la posicion va con signo
El reporte en vivo fue «la nave se para en seco en el borde» — por el lado
del 0. No era el clamp de radiacion: eran cinco capas que daban por hecho
coordenadas sin signo, cada una suficiente sola para dejar el borde en pared:
el Vector2.ZERO del cliente, el uint del cable, el Math.Max(0, v) de
WireMapping.Round, el 0 de OnMoveIntent y el INT UNSIGNED de la BD.

Aqui caen tres: OnMoveIntent clampea a [-RadiationMargin, Bounds+Margin],
Round ya no recorta al 0, y PlayerData.PosX/PosY, SaveShipState y
MoveIntentCmd van con signo. Requiere el protocolo con sint y la migracion
2026.09.01.1 — los tres se despliegan juntos.

Pruebas: el clamp del lado cercano es -1000 y no 0, y cruzar el 0 cobra el
10% igual que cruzar el limite lejano (100/100).

+++ src/MexOrbit.GameServer.Application/Commands.cs
@@ -23,7 +23,7 @@ public sealed record ResumeCmd(IClientPort Port, long AccountId, long SessionId,
 
 public sealed record LeaveCmd(IClientPort Port, string Reason) : WorldCmd(Port);
 
-public sealed record MoveIntentCmd(IClientPort Port, ulong Seq, uint TargetX, uint TargetY)
+public sealed record MoveIntentCmd(IClientPort Port, ulong Seq, long TargetX, long TargetY)
     : WorldCmd(Port);
 
 public sealed record PongCmd(IClientPort Port, ulong Nonce) : WorldCmd(Port);
